Rename the module "utility.search_mode_utilities" to "utility.search"
